Find spot
This is just a file with the find spot method. This method was mostly working in its previous states, but I tried to make it work 10/10 times with this one. It's not the best solution or anything, but it seems to be working for me, so before I mess things up again like yesterday I am just gonna upload this now as it is, incase we want to use it.