A find middle method that can be used with park
it doesnt always get you the exact middle, but that is quite an impossible feat, so I think this is definitely good enough. The car often ends up in the middle and if its a bit slanted to one side its only by a few centimeters so I think it's good!