ObjectDetection with KP for PI
This is what I was working on, object detection with KeyPoints, we tried it on and it worked on the webcamera but when I tried it on the PI it didn't work well at all. So I do not think we will pursue this code any further, but here it is anyhow for documentation purposes. 
I think this is the latest py file I was working on for this code, but I am not sure, anyway it should look something like this.