SIFT Object detection updated for the pi camera
This was the code I was working on using OpenCV and sift for object detection. Here I tried to make a version for the Picamera, but I couldn't get it to work. Maybe if I had used imutils here instead of the picamerarray it could've worked, but there was no time to try.