Object detection using SIFT, OpenCV, keypoints
This is the original code I was working on with the SIFT stuff. I used this code with my webcamera and it could detect objects very nicely! Different lighting conditions made it more difficult for it to work though, since it only based matches upon 1 single sample image.
Anyway, I tried to make this code work for the PIcamera (uploaded that version too), but it didnt, so in the end we decided to use another method for object detection instead.